Pulled CLI tools into a single crate and consolidated.
Moved the CLI tools into a single crate to reduce dependencies overall. Also exploded status into a cmdline tool that can address functionality individually from the cmdline. This is now a swiss army knife to talking to terminals.

The update should be eventually incorporated into this one, but I ran out of time for this today.